README updates for semantics, and reorganizing of the semantic tag types.
